Functioning On Server Model
